Start the type review: the indices that only index
From 241 unique sign-conversion locations to 167. The raw warning count
was 606, because a header is counted once per translation unit - the same
lesson clang-tidy taught in phase 8.

Converted: the boundary engine's index parameter, the pure loop indices of
the five algorithms and the engines, and the four loops of the 1D union
along with their bucketSize.

Those four deserve a note, because they are the case where the sign has
teeth: they read bucket[j - 1], and an unsigned j of 0 would become a very
large number. They are safe, BUT ONLY BY REASONING - the one and two point
cases are handled in earlier branches, so bucketSize is at least three
there, and stepping by two the else branch is only reached with j at two
or more. This is exactly why the item cannot be a regex.

And a correction to this section's rule: there are not two kinds of
integer here, there are three. Indices, which only index, and become
size_t. Counts, which are validated input doing arithmetic, and stay
signed. And tagged selectors, which I had not seen: algorithm MC's
"parameter" is 0 for the gain, 1..nz for a zero and above that for a pole,
with an "if (parameter < 0)" where the negative means something. That one
stays signed, and its conversions when indexing are deliberate boundary
crossings rather than noise to paper over.

+++ src/core/boundaries/boundary_engine.cpp
@@ -198,7 +198,7 @@ std::vector <double> * BoundaryEngine::omega(){
 void BoundaryEngine::traceFrequency(double omega, std::map<std::string, TraceSet> & bound,
                                     const BoundarySheets & sheets,
                                     std::map<std::string, TraceLabels> & traceMetadata,
-                                    complex <double> p0, const ComplexCloud & valueSet, std::int32_t index,
+                                    complex <double> p0, const ComplexCloud & valueSet, std::size_t index,
                                     double phaseSpan, double magnitudeSpan, double phaseBottom, double magnitudeBottom){
 
 
@@ -263,7 +263,7 @@ void BoundaryEngine::traceFrequency(double omega, std::map<std::string, TraceSet
 void BoundaryEngine::traceFrequency(double omega, std::map<std::string, TraceSet> & bound,
                                     const BoundarySheetsCuda & cudaSheets,
                                     std::map<std::string, TraceLabels> & traceMetadata,
-                                    complex <double> p0, const ComplexCloud & valueSet, std::int32_t index,
+                                    complex <double> p0, const ComplexCloud & valueSet, std::size_t index,
                                     double phaseSpan, double magnitudeSpan, double phaseBottom, double magnitudeBottom){
 
     if (m_trackingMask.at(index)){
@@ -559,7 +559,7 @@ double violatingDb(double valueDb)
 void BoundaryEngine::computeFrequency (double omega, LtiSystem * plant,
                                        const ComplexCloud & valueSet,
                                        const std::vector <double> & phases,
-                                       const std::vector <double> & magnitudes, std::int32_t index){
+                                       const std::vector <double> & magnitudes, std::size_t index){
 
     //Nominal plant at this design frequency.
     complex <double> p0 = plant->evaluate(omega);
